% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.r
\name{bind}
\alias{bind}
\title{Binds arrays together disregarding names}
\usage{
bind(arrayList, along = length(dim(arrayList[[1]])) + 1)
}
\arguments{
\item{arrayList}{A list of n-dimensional arrays}

\item{along}{Along which axis to bind them together; default: new axis}
}
\value{
A joined array
}
\description{
Binds arrays together disregarding names
}
