% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.r
\name{collect}
\alias{collect}
\title{Converts a logical matrix to a list of character vectors}
\usage{
collect(x, along = 2)
}
\arguments{
\item{x}{A logical matrix}

\item{along}{Which axis to spread mask on}
}
\value{
A character vector or list thereof
}
\description{
This currently only supports x with only one non-zero element
}
