% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.r
\name{construct}
\alias{construct}
\title{A wrapper around reshape2::acast using a more intuitive formula syntax}
\usage{
construct(formula, data, fill = NULL, fun.aggregate = NULL, ...)
}
\arguments{
\item{formula}{A formula: value [+ value2 ..] ~ axis1 [+ axis2 + axis n ..]}

\item{data}{A data frame (TODO: handle envs, NULL, etc.)}

\item{fill}{Value to fill array with if undefined}

\item{fun.aggregate}{Function to aggregate multiple values for the same position; default: error}

\item{...}{Additional arguments passed to reshape2::acast}
}
\value{
A structured array
}
\description{
A wrapper around reshape2::acast using a more intuitive formula syntax
}

