% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimnames.r
\name{dimnames}
\alias{dimnames}
\title{Return dimension names of an array respecting the number of dimensions}
\usage{
dimnames(x, along = TRUE, null_as_integer = FALSE, drop = TRUE)
}
\arguments{
\item{x}{An n-dimensional array}

\item{along}{Limit to dimension (default: all)}

\item{null_as_integer}{Whether nameless dimensions should be \code{NULL} or numbered}

\item{drop}{Drop list of only one axis requested (default: TRUE)}
}
\value{
A list of dimension names with length \code{length(ndim(X))}
}
\description{
Act on each element if 'x' is a list
}

