% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{drop_if}
\alias{drop_if}
\title{Drop unused dims if flag is TRUE}
\usage{
drop_if(x, flag)
}
\arguments{
\item{x}{An array object}

\item{flag}{Whether to drop unused dimensions}
}
\value{
The object in full or with dropped dimensions
}
\description{
Drop unused dims if flag is TRUE
}
