% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{duplicated}
\alias{duplicated}
\title{duplicated() function with extended functionality}
\usage{
duplicated(x, ..., all = F, random = F)
}
\arguments{
\item{x}{Object to check for duplicates on}

\item{...}{Arguments to be passed to R's `duplicated()`}

\item{all}{Return all instances of duplicated entries}

\item{random}{Randomly select with entry is marked as duplicate}
}
\description{
duplicated() function with extended functionality
}

