% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{elementnames}
\alias{elementnames}
\title{Assigns matrix element names by row- and column names}
\usage{
elementnames(X, sep = ":")
}
\arguments{
\item{X}{An n-dimensional array}

\item{sep}{Character to use when assembling names for each element}
}
\value{
A vector of name characters for each element of the array \code{X}
}
\description{
Assigns matrix element names by row- and column names
}

