% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{expand_grid}
\alias{expand_grid}
\title{expand.grid() function not converting to factors}
\usage{
expand_grid(..., KEEP.OUT.ATTRS = FALSE, stringsAsFactors = FALSE)
}
\description{
expand.grid() function not converting to factors
}

