% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.r
\name{flatten}
\alias{flatten}
\title{Flattens an array along an axis}
\usage{
flatten(X, along, name_sep = NA)
}
\arguments{
\item{X}{Array}

\item{along}{Along which axis to bind them together}

\item{name_sep}{Which character to use for naming new arrays [default: NA, do not touch names]}
}
\value{
An array with n-1 dimensions
}
\description{
Flattens an array along an axis
}
