% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{\%or\%}
\alias{\%or\%}
\title{Operator for array-like logical operations}
\usage{
a \%or\% b
}
\arguments{
\item{a}{First vector}

\item{b}{Second vector}
}
\value{
TRUE/FALSE for each element
}
\description{
Operator for array-like logical operations
}
