% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{guess_structure}
\alias{guess_structure}
\title{Infer array structure from data.frame}
\usage{
guess_structure(df, verbose = TRUE)
}
\arguments{
\item{df}{A data.frame with ordered axes, value field last}

\item{verbose}{Print message with inferred structure (default: TRUE)}
}
\value{
A formula describing this structure
}
\description{
Infer array structure from data.frame
}
