% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r, R/intersect.r
\name{intersect}
\alias{intersect}
\title{intersect() function that takes an arbitrary number of elements}
\usage{
intersect(..., along = 1, envir = parent.frame(), drop = FALSE)

intersect(..., along = 1, envir = parent.frame(), drop = FALSE)
}
\arguments{
\item{...}{Arbitrary elements that `base::intersect` should be called on}

\item{along}{The axis along which to intersect}

\item{envir}{A list or environment to act upon}

\item{...}{Arrays that should be intersected}
}
\description{
intersect() function that takes an arbitrary number of elements

TODO: accept along=c(1,2,1,1...) [maybe list w/ vectors as well?]
TODO: accept data=env/list arg? [sig-comb/drug-tissue/assocs.r#62-65]
}

