% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect.r
\name{intersect_list}
\alias{intersect_list}
\title{Intersects a lits of arrays for common dimension names}
\usage{
intersect_list(l., along = 1, drop = FALSE)
}
\description{
Intersects a lits of arrays for common dimension names
}

