% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.r
\name{lambda}
\alias{lambda}
\title{Lambda syntax for array iteration}
\usage{
lambda(fml, along, group = c(), simplify = TRUE, expand_grid = TRUE,
  envir = parent.frame())
}
\arguments{
\item{fml}{A call prefixed with a tilde}

\item{along}{A named vector which objects to subset (eg: c(x=1))}

\item{group}{Not implemented}

\item{simplify}{Return array instead of index+result if scalar}

\item{expand_grid}{Use all combinations of indices (default: TRUE)}

\item{envir}{Environment where variables can be found}
}
\description{
Lambda syntax for array iteration
}
