% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.r
\name{mask}
\alias{mask}
\title{Converts a list of character vectors to a logical matrix}
\usage{
mask(x)
}
\arguments{
\item{x}{A list of character vectors}
}
\value{
A logical occurrence matrix
}
\description{
Converts a list of character vectors to a logical matrix
}

