% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.r
\name{match}
\alias{match}
\title{match() function with extended functionality}
\usage{
match(x, from, to, filter_from = NULL, filter_to = NULL,
  data = parent.frame(), fuzzy_level = 0, table = FALSE, na_rm = FALSE,
  warn = !table && fuzzy_level > 0)
}
\arguments{
\item{x}{Vector of identifiers that should be mapped}

\item{from}{Vector of identifiers that can be mapped}

\item{to}{Matched mapping for all identifiers}

\item{filter_from}{Restrict matching to a subset from `from`}

\item{filter_to}{Restrict matching to a subset from `to`}

\item{data}{List containing the data `from` and `to` reference}

\item{fuzzy_level}{0 for exact, 1 punctuation, and 2 closest character}

\item{table}{Return a matching table instead of just the matches}

\item{na_rm}{Flag to remove items that can not be mapped}

\item{warn}{Display warning for all fuzzy matches}
}
\value{
Mapped values
}
\description{
match() function with extended functionality
}

