% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{named_dots}
\alias{named_dots}
\title{Return a list of named dot-arguments}
\usage{
named_dots(...)
}
\arguments{
\item{...}{Function arguments}
}
\value{
Named function arguments
}
\description{
Return a list of named dot-arguments
}
