% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narray-package.r
\docType{package}
\name{narray}
\alias{narray}
\alias{narray-package}
\title{R package for subset- and name-aware array utility functions}
\description{
Stacking arrays according to dimension names, subset-aware splitting and
mapping of functions, intersecting along arbitrary dimensions, converting to
and from data.frames, and many other helper functions.
}
