% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{pb}
\alias{pb}
\title{Progress bar format to be consistent}
\usage{
pb(ticks)
}
\arguments{
\item{ticks}{Number of ticks the bar has}
}
\value{
A progress bar object
}
\description{
Progress bar format to be consistent
}
