% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.r
\name{rep}
\alias{crep}
\alias{rep}
\alias{rrep}
\title{Repeats an array along an arbitrary axis}
\usage{
rep(x, n, along = 1)

crep(x, n)

rrep(x, n)
}
\arguments{
\item{x}{An array object}

\item{n}{Integer, how often to repeat}

\item{along}{Along which axis to repeat (default: 1)}
}
\value{
An array that is repeated `n` times on axis `along`
}
\description{
Repeats an array along an arbitrary axis
}

