% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{restore_null_dimnames}
\alias{restore_null_dimnames}
\title{If no dimnames, return NULL and not list of NULLs}
\usage{
restore_null_dimnames(x)
}
\arguments{
\item{x}{An array object}
}
\value{
The object with NULL if no dimnames
}
\description{
If no dimnames, return NULL and not list of NULLs
}
