% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.r
\name{split}
\alias{split}
\title{Splits and array along a given axis, either totally or only subsets}
\usage{
split(X, along, subsets = c(1:dim(X)[along]), drop = NULL)
}
\arguments{
\item{X}{An array that should be split}

\item{along}{Along which axis to split; use -1 for highest dimension}

\item{subsets}{Whether to split each element or keep some together}

\item{drop}{Remove unused dimensions after mapping
default: drop if all resulting arrays have same number of dimensions}
}
\value{
A list of arrays that combined make up the input array
}
\description{
Splits and array along a given axis, either totally or only subsets
}
