% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.r
\name{summarize}
\alias{summarize}
\title{Summarize a matrix analogous to a grouped df in dplyr}
\usage{
summarize(x, along = 1, to, from = dimnames(x)[[along]], ..., FUN)
}
\arguments{
\item{x}{A matrix}

\item{along}{Along which axis to summarize}

\item{to}{Names that this dimension should be summarized to}

\item{from}{Names that match the dimension `along`}

\item{...}{Parameters passed to `base/match`}

\item{FUN}{Which function to apply, default is throwing error on aggregation}
}
\value{
A summarized matrix as defined by `from`, `to`
}
\description{
Summarize a matrix analogous to a grouped df in dplyr
}
