% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.r
\name{translate}
\alias{translate}
\title{Translate an axis between two sets of identifiers}
\usage{
translate(x, along = 1, to, from = dimnames(x)[[along]], ..., FUN,
  na_rm = FALSE)
}
\arguments{
\item{x}{A matrix}

\item{along}{Along which axis to summarize}

\item{to}{Names that this dimension should be summarized to}

\item{from}{Names that match the dimension `along`}

\item{...}{Parameters passed to `match`}

\item{FUN}{Which function to apply, default is throwing error on aggregation}

\item{na_rm}{Remove values that were translated to NAs}
}
\value{
A summarized matrix as defined by `from`, `to`
}
\description{
Translate an axis between two sets of identifiers
}
