% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors_to_row_or_col.r
\name{vectors_to_row_or_col}
\alias{vectors_to_row_or_col}
\title{Converts vectors in a list to row- or column vectors}
\usage{
vectors_to_row_or_col(xlist, along)
}
\arguments{
\item{xlist}{List of array-like elements and vectors}

\item{along}{Along which dimension vectors should be aligned}
}
\value{
List where vectors are replaced by row- or col vectors (2d)
}
\description{
Converts vectors in a list to row- or column vectors
}
