% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{which}
\alias{which}
\title{A multidimensional \code{which} function}
\usage{
which(A)
}
\arguments{
\item{A}{N-dimensional logical array}
}
\value{
A matrix with indices where \code{A == TRUE}
}
\description{
From: \url{http://www.r-bloggers.com/a-multidimensional-which-function/}
}

