% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.r
\name{bind}
\alias{bind}
\title{Binds arrays together disregarding names}
\usage{
bind(..., along = length(dim(arrayList[[1]])) + 1)
}
\arguments{
\item{...}{N-dimensional arrays, or a list thereof}

\item{along}{Along which axis to bind them together (default: new axis)}
}
\value{
A joined array
}
\description{
Binds arrays together disregarding names
}
