% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.r
\name{construct}
\alias{construct}
\title{A wrapper around reshape2::acast using a more intuitive formula syntax}
\usage{
construct(data, formula = guess_structure(data), fill = NA,
  name_axes = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{formula}{A formula: value ~ axis1 [+ axis2 + axis n ..]}

\item{fill}{Value to fill array with if undefined}

\item{name_axes}{Keep column names of `data` as axis names}
}
\value{
A structured array
}
\description{
The construct() function can be called either with the data.frame as the
first argument or the formula and then specify `data=<data.frame>`
}
