% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.r
\name{construct}
\alias{construct}
\title{A wrapper around reshape2::acast using a more intuitive formula syntax}
\usage{
construct(data, formula, fill = NULL, fun.aggregate = NULL, ...)
}
\arguments{
\item{data}{A data frame}

\item{formula}{A formula: value ~ axis1 [+ axis2 + axis n ..]}

\item{fill}{Value to fill array with if undefined}

\item{fun.aggregate}{Function to aggregate multiple values for the same position; default: error}

\item{...}{Additional arguments passed to reshape2::acast}
}
\value{
A structured array
}
\description{
The construct() function can be called either with the data.frame as the
first argument or the formula and then specify `data=<data.frame>`
}
