% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.r
\name{\%or\%}
\alias{\%or\%}
\title{operator for array-like logical ops}
\usage{
a \%or\% b
}
\arguments{
\item{a}{First vector}

\item{b}{Second vector}
}
\value{
TRUE/FALSE for each element
}
\description{
operator for array-like logical ops
}

