% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect.r
\name{intersect_list}
\alias{intersect_list}
\title{Intersects a lits of arrays for common dimension names}
\usage{
intersect_list(l., along = 1, drop = FALSE)
}
\arguments{
\item{l.}{List of arrays to perform operations on}

\item{along}{The axis along which to intersect}

\item{drop}{Drop unused dimensions on result}
}
\description{
Intersects a lits of arrays for common dimension names
}

