% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_list.r
\name{intersect_list}
\alias{intersect_list}
\title{Intersects a lits of arrays for common dimension names}
\usage{
intersect_list(l., along = 1, drop = FALSE, fail_if_empty = TRUE)
}
\arguments{
\item{l.}{List of arrays to perform operations on}

\item{along}{The axis along which to intersect}

\item{drop}{Drop unused dimensions on result}

\item{fail_if_empty}{Stop if intersection yields empty set}
}
\description{
Intersects a lits of arrays for common dimension names
}
