% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like.r
\name{like}
\alias{like}
\title{Reshapes \code{x} to be like \code{like}, including dimension names}
\usage{
like(x, like)
}
\arguments{
\item{x}{An n-dimensional array}

\item{like}{An n-dimensional array whose form \code{X} should inherit}
}
\value{
An array with values of \code{X} and structure of \code{like}
}
\description{
Reshapes \code{x} to be like \code{like}, including dimension names
}

