% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_simple.r
\name{map_simple}
\alias{map_simple}
\title{Apply function that preserves order of dimensions}
\usage{
map_simple(X, along, FUN, drop = TRUE, ...)
}
\arguments{
\item{X}{An n-dimensional array}

\item{along}{Along which axis to apply the function}

\item{FUN}{A function that maps a vector to the same length or a scalar}

\item{drop}{Remove unused dimensions after mapping; default: TRUE}

\item{...}{Arguments passed to the function}
}
\value{
An array where \code{FUN} has been applied
}
\description{
Apply function that preserves order of dimensions
}
