% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.r
\name{mask}
\alias{mask}
\title{Converts a list of character vectors to a logical matrix}
\usage{
mask(x, along = 2, na_rm = FALSE)
}
\arguments{
\item{x}{A list of character vectors}

\item{along}{Which axis to spread mask on}

\item{na_rm}{Remove values that were translated to NAs}
}
\value{
A logical occurrence matrix
}
\description{
Converts a list of character vectors to a logical matrix
}
