% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt.r
\name{melt}
\alias{melt}
\title{Function to melt data.frame from one or multiple arrays}
\usage{
melt(..., dimnames = NULL, na_rm = TRUE)
}
\arguments{
\item{...}{Array[s] or data.frame[s] to be melted}

\item{dimnames}{List of names along the dimensions}

\item{na_rm}{Remove rows with NAs}
}
\value{
data.frame with 'value' (or object names if multiple) indexed by axes
}
\description{
Function to melt data.frame from one or multiple arrays
}
