% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.r
\name{stack}
\alias{stack}
\title{Stacks arrays while respecting names in each dimension}
\usage{
stack(arrayList, along = length(dim(arrayList[[1]])) + 1, fill = NA,
  drop = FALSE)
}
\arguments{
\item{arrayList}{A list of n-dimensional arrays}

\item{along}{Which axis arrays should be stacked on (default: new axis)}

\item{fill}{Value for unknown values (default: \code{NA})}

\item{drop}{Remove unused dimensions after mapping; default: TRUE}
}
\value{
A stacked array, either n or n+1 dimensional
}
\description{
Stacks arrays while respecting names in each dimension
}

