% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.r
\name{which}
\alias{which}
\title{A multidimensional \code{which} function}
\usage{
which(x, drop = TRUE)
}
\arguments{
\item{x}{N-dimensional logical array}

\item{drop}{Return a vector if called on a vector}
}
\value{
A matrix with indices where \code{A == TRUE}
}
\description{
A multidimensional \code{which} function
}

